✨ feat: Add Jetpack SourcePost autocmd to initialize plugin on opening Vim
This commit adds an autocmd to the Jetpack plugin's end function, which calls the jetpack#init() function after sourcing $MYVIMRC. The jetpack#init() function contains necessary initialization steps for the plugin, and is triggered by the Jetpack SourcePost event. This change ensures that the Jetpack plugin is properly initialized on opening Vim.