✨ feat: Add support for plugin dependencies in Jetpack
This commit adds support to the Jetpack Vim plugin manager to handle plugin dependencies declared with the `requires` attribute in the plugin configuration. The new implementation recursively installs all required plugins and their dependencies. In this specific case, the change consists of adding a `for` loop that invokes the `use` function on each element of the `plugin.requires` table.