♻️ refactor: simplify plugin.requires handling in use()
This commit refactors the `use()` function in `jetpack.vim` to simplify the handling of `plugin.requires`. Instead of looping through the list of requirements and checking their type, this commit uses a conditional statement to determine whether each requirement is a string or an object. If the requirement is a string, it is added to `plugin.requires` as is. If it is an object, its `as` property is used, or its first value if `as` is not defined. This change makes the code more concise and easier to understand.