✨ feat: Add support for multiple plugin dependencies
This commit adds support for specifying multiple plugin dependencies in the Jetpack Vim plugin. Previously, only a single dependency could be specified as a string, but now an array of strings can be used to specify multiple dependencies.

The code has been updated to automatically convert a single string dependency to an array with a single element, so that the rest of the code can treat all dependencies uniformly regardless of whether there is one or many.

Additionally, this commit fixes a bug where the 'as' field was not correctly set when a dependency was specified as an object with an 'as' field. Now the 'as' field is correctly extracted and set as the dependency name.

This should make it easier for users of the Jetpack plugin to specify and manage their plugin dependencies.