✨ feat: Add support for loading Jetpack packages before Vim startup
This commit adds support for loading Jetpack packages before Vim startup by adding the necessary code in the `jetpack#load()` function. Specifically, if Vim did not enter (i.e., it is starting up), the function will modify the `&rtp` variable to include the package's runtimepath and the `after/` directory. Additionally, it will define two autocmds (`JetpackPre:init` and `JetpackPost:init`) that will be executed before and after the package is loaded, respectively. This feature allows Jetpack packages to be loaded early, ensuring they are available when Vim starts up.